+++ db/block.json
@@ -1,1 +1,4 @@
-{"last_processsed_block":226449,"twitterRefreshToken":"SXdiVTZra0kxaEtQa2pqMkNXLTBVMkEtSmctbEx5bmZfQ0Q1QXluNlBHelNmOjE2OTUxMjYxNDYxNTM6MToxOnJ0OjE"}
+{
+  "last_processsed_block": 226449,
+  "twitterRefreshToken": "RWx1NEZMbzVxOEFTaDM3dFUzMVd5akJyb0RHQmhKM3BDc25XTjRJZjVKTEV2OjE2OTUyMDU2OTc1NjY6MToxOnJ0OjE"
+}
